requirements.txt
# venvg, python -V => Python 3.12.8
# start with https://github.com/facebookresearch/xformers
# for cuda 12.4 (for other cudas, change tail) or 12.8
# pip install -U xformers --index-url https://download.pytorch.org/whl/cu124
# torch - this is installed by the xformers - install rest below
# numpy

torchvision
opencv-python
pillow
matplotlib
requests
diffusers
transformers
ftfy
gradio
git+https://github.com/openai/CLIP.git
huggingface-hub
accelerate
peft

# 31-03-2025 14:24
# install clipseg
# ./install.sh
